New Function to place disks and mouse click to place disk
Left original code in incase I am not understanding parts of it and shouldnt have changed it. New function just takes a color and board cords to place the disk. Mouse event checks for clicks on the board and converts the coordinates to board coordinates then places a disk there. Not made to interact with anything yet.