Added some move checks
The adjacent piece check looks pretty gross but it works. I changed over to an int array because I couldn't see how we would handle the line checks with objects.

Right now the array just has 0 for no disk, 1 for a white disk, 2 for a black disk, and then a -1 for a valid move.

If you try to make an invalid move the array will print itself to the console for debug purposes right now.